refactor: clean up ng file processor and add snapshot tests
